If filesystem is read-only, we have to copy the content out instead of mv.
For example, the tar.gz filesystem is read-only, so we copyFrom out of it.
Also, copy works such that you have to create a "dummy" file that we write
the contents to, which is inside of destDirectory. Otherwise if you just try
to copyFrom from:src and to:destDir then you end up writing to a file which
is named what your destDirectory should have been.